requirements.txt
# Editable install with no version control (aclpubcheck==0.1)
-e /data/guowenyu/aclpubcheck-main
antlr4-python3-runtime==4.8
bibtexparser==1.4.1
bitarray==2.8.0
certifi==2019.11.28
cffi==1.15.1
chardet==5.0.0
click==8.0.4
colorama==0.4.5
cryptography==40.0.2
Cython==0.29.35
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958435052/work
-e git+https://github.com/zhulifengsheng/fairseq_mmt.git@4f7bb90f6b98e71b41690c24caae75baff9513d2#egg=fairseq
hydra-core==1.0.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
latexcodec==2.0.1
lxml==4.9.2
mkl-fft==1.2.0
mkl-random==1.2.0
mkl-service==2.3.0
numpy==1.19.5
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
omegaconf==2.0.6
opencv-python==4.4.0.46
packaging==21.3
pandas==1.1.5
pdfminer.six==20211012
pdfplumber==0.6.0
Pillow==8.4.0
portalocker==2.0.0
pybtex==0.24.0
pycparser==2.21
pylatexenc==2.10
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0
rebiber==1.1.3
regex==2023.6.3
sacrebleu==2.3.1
sacremoses==0.0.53
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tabulate==0.8.10
termcolor==1.1.0
timm==0.4.12
torch==1.10.1
torchaudio==0.10.1
torchvision==0.8.0
tqdm==4.64.1
tsv==1.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1644850595256/work
Unidecode==1.3.7
Wand==0.6.11
zipp==3.6.0
